Dispose 2D images.
Textures for the 2D images are disposed upon exiting them,
and they are also disposed if, upon loading the texture, a different
image has already been selected, or the texture for this image has
already separately been loaded (if the user was spam-clicking images).

Camera movement logic when going in and out of the image was also moved.

The camera now returns to the previous position and pivot upon exiting
instead of the default position and pivot, to match 360Images. But both
do not preserve the camera direction, instead pointing the camera
directly at the pivot.

This change depends on another pull request in the cpms-web repo.