Fix TDImage alignment with pointclouds
Now, if a matched pair of TDImages and pointclouds have the same
position, they will be perfectly aligned with each other.

This works by setting pointcloud's offset variable to (0,0,0) to match
the TDImages. It would be better if the TDImage was given an offset to
match the pointcloud, but there is currently no link between a TDImage
and its matching pointcloud to decide which one should take the offset
of which.

This change also messes with pointclouds that were manually aligned to
the TDImages, because now they're being double-moved and no longer
match, so they will have to be fixed. If you just reset the pointcloud
and TDImage positions to 0, they will line up.