Navigate in 360images by clicking small objects
When focused on a 360image, shrunken versions of other 360images are
visible and clickable to switch focus to those.

360images from other sets are now visible and not shrunken. They are
clickable too, to switch focus directly to another 360image set. The
user can also now use the mini scene to switch focus directly to another
360image set as well.

I wrote a commented version of the code in Images360.js that makes the
spheres have the same screen-size no matter how far away they are, but
it seemed less intuitive to understand where I was looking at, so I did
not use this option. It also may be more computationally expensive
because it requires a distance calculation for each 360image.

Since other 360image sets are not shrunken, two overlapping 360image
sets may be hard to navigate. Two overlapping 360images from the same
set can also be hard to navigate, because when you are inside one, you
can only click the other image, so you can only go between the two and
can't click other images. This can only be escaped by unfocusing or
using the mini-scene.

Sometimes a 360image is so close that it passes partially in front of
the near clipping plane, making it look strange, or it may be entirely
in front of the near clipping plane, so you can't see it but it still
catches your clicks and hovers.

This commit depends on a commit in cpms-web.