Fixed some 360image bugs.
Switching back to the main scene from another view such as crosssection
no longer breaks 360images.

Changing the time on the scheduler now changes the 360image that appears
in the miniscene according to how it looks in the main scene.

The miniscene now disappears when it is not currently showing anything.
This also means it disappears in other views where it is not being used.

This commit depends on a commit to cpms-web.