Allow OrbitControls to be restricted to 2D rotate.
Change boolean allowRotation to string allowedRotation for OrbitControls

allowedRotation = "all" is normal. allowedRotation = "x" or "y" or "z"
restricts its rotation to the "x" or "y" or "z" dimensions.

But since it is based on intrinsic angles, it only works if the camera
is oriented a certain way. Making it work for all orientations might be
messier, and is not necessary.