Allow input listeners on mini canvases
Issues: the listener fires when dragging to resize.
I wasn't able to find a way to stop this via the libraries or css. I
think it would be easiest to turn go in PotreeViewer.component.js, and
add start and end listeners that toggle a flag which makes
InputHandler.js's listeners do nothing.

A red error pops up from Engine.components.js on line 171 because
that.bimHighlighter.selection.ghost is undefined. Depending on whether
you click the main canvas or the minicanvas, you get a different "that"
and "that.bimHighlighter". I don't know if this is caused by my commit
or not.